PROJECT3/requirements.txt
python-telegram-bot==13.15
openai>=1.0.0
python-dotenv
loguru